v0.1.29: extract music handlers, fix CI duplicate-channel detection
- Move load_demo_tracks and generate_music out of chat.js into new src/ipc/music.js
- Remove duplicate ipcMain.handle registrations that short-circuited the registerAll chain and broke settings/npc/filesystem handlers
- Harden CI: drop "|| true" masks, syntax-check all ipc/*.js, add registerAll smoke test with stubbed ipcMain that fails on duplicate channels
- Bump version to 0.1.29
