+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "incognide",
-  "version": "0.1.28",
+  "version": "0.1.29",
   "description": "Explore the unknown, build the future, own your data.",
   "author": "Chris Agostino <info@npcworldwi.de>",
   "main": "src/main.js",
